pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Garkusha Valery <aston585v@gmail.com>"]
readme = "README.md"
packages = [{include = "gendiff"}]
repository = "https://github.com/Aston585/python-project-50.git"
classifiers = [
    "Programming Language :: Python :: 3",
    "Environment :: Console",
    "License :: Freeware",
]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"
flake8 = "^6.0.0"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
